refactor(auth-lib): move useAuthType from lib/hooks.ts to lib/auth/hooks.ts
