+++ web/src/sections/actions/modals/MCPAuthenticationModal.tsx
@@ -36,7 +36,7 @@ import { PerUserAuthConfig } from "@/sections/actions/PerUserAuthConfig";
 import { updateMCPServerStatus, upsertMCPServer } from "@/lib/tools/mcpService";
 import { toast } from "@/hooks/useToast";
 import { SvgArrowExchange } from "@opal/icons";
-import { useAuthType } from "@/lib/hooks";
+import { useAuthType } from "@/lib/auth/hooks";
 import { AuthType } from "@/lib/auth/types";
 
 interface MCPAuthenticationModalProps {
