+++ web/src/sections/actions/modals/OpenAPIAuthenticationModal.tsx
@@ -17,7 +17,7 @@ import KeyValueInput, {
 import { OAuthConfig } from "@/lib/tools/interfaces";
 import { getOAuthConfig } from "@/lib/oauth/api";
 import { SvgArrowExchange } from "@opal/icons";
-import { useAuthType } from "@/lib/hooks";
+import { useAuthType } from "@/lib/auth/hooks";
 import { AuthType } from "@/lib/auth/types";
 
 export type AuthMethod = "oauth" | "custom-header" | "pt-oauth";
